% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floyd_steinberg_dithering.R
\name{floyd_dither}
\alias{floyd_dither}
\title{Floyd dither}
\usage{
floyd_dither(path)
}
\arguments{
\item{path}{A path to an image}
}
\value{
An image object
}
\description{
Floyd dither
}
\examples{
floyd_dither(path_to_file("dumbo.jpg"))
}
