% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{path_to_file}
\alias{path_to_file}
\title{Title}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{A path to a file}
}
\value{
A path to a file
}
\description{
Title
}
\examples{
path_to_file("dumbo.jpg")
}
