% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floyd_steinberg_dithering.R
\name{write_image}
\alias{write_image}
\title{Write image}
\usage{
write_image(image, path)
}
\arguments{
\item{image}{An image object}

\item{path}{A path to save the image to}
}
\description{
Write image
}
\examples{
image <- magick::image_read(path_to_file("dumbo.jpg"))
write_image(image, "dumbo.jpg")
}
